extracted PooledLintFileTask/PooledLintSourceTask which need to be autoloadable classes
